+++ cli/cli.go
@@ -52,11 +52,17 @@ func init() {
 	fs.StringVar(&defaults.SSL.CertPath, "ssl_cert", "", "")
 	fs.StringVar(&defaults.SSL.KeyPath, "ssl_key", "", "")
 
+	fs.StringVar(&defaults.BroadcastAdapter, "broadcast_adapter", "redis", "")
+
 	fs.StringVar(&defaults.Redis.URL, "redis_url", redisDefault, "")
 	fs.StringVar(&defaults.Redis.Channel, "redis_channel", "__anycable__", "")
 	fs.StringVar(&defaults.Redis.Sentinels, "redis_sentinels", "", "")
 	fs.IntVar(&defaults.Redis.SentinelDiscoveryInterval, "redis_sentinel_discovery_interval", 30, "")
 	fs.IntVar(&defaults.Redis.KeepalivePingInterval, "redis_keepalive_interval", 30, "")
+
+	fs.IntVar(&defaults.HTTPPubSub.Port, "http_broadcast_port", 8090, "")
+	fs.StringVar(&defaults.HTTPPubSub.Path, "http_broadcast_path", "/_broadcast", "")
+	fs.StringVar(&defaults.HTTPPubSub.Secret, "http_broadcast_secret", "", "")
 
 	fs.StringVar(&defaults.RPC.Host, "rpc_host", "localhost:50051", "")
 	fs.IntVar(&defaults.RPC.Concurrency, "rpc_concurrency", 28, "")
@@ -123,11 +129,17 @@ OPTIONS
   --ssl_cert                             SSL certificate path, env: ANYCABLE_SSL_CERT
   --ssl_key                              SSL private key path, env: ANYCABLE_SSL_KEY
 
+  --broadcast_adapter                    Broadcasting adapter to use (redis or http), default: redis, env: ANYCABLE_BROADCAST_ADAPTER
+
   --redis_url                            Redis url, default: redis://localhost:6379/5, env: ANYCABLE_REDIS_URL, REDIS_URL
   --redis_channel                        Redis channel for broadcasts, default: __anycable__, env: ANYCABLE_REDIS_CHANNEL
   --redis_sentinels                      Comma separated list of sentinel hosts, format: 'hostname:port,..', env: ANYCABLE_REDIS_SENTINELS
   --redis_sentinel_discovery_interval    Interval to rediscover sentinels in seconds, default: 30, env: ANYCABLE_REDIS_SENTINEL_DISCOVERY_INTERVAL
   --redis_keeepalive_interval            Interval to periodically ping Redis to make sure it's alive, default: 30, env: ANYCABLE_REDIS_KEEPALIVE_INTERVAL
+
+  --http_broadcast_port                  HTTP pub/sub server port, default: 8090, env: ANYCABLE_HTTP_BROADCAST_PORT
+  --http_broadcast_path                  HTTP pub/sub endpoint path, default: /_broadcast, env: ANYCABLE_HTTP_BROADCAST_PATH
+  --http_broadcast_secret                HTTP pub/sub authorization secret, default: "" (disabled), env: ANYCABLE_HTTP_BROADCAST_SECRET
 
   --rpc_host                             RPC service address, default: localhost:50051, env: ANYCABLE_RPC_HOST
   --rpc_concurrency                      Max number of concurrent RPC request; should be slightly less than the RPC server concurrency, default: 28, env: ANYCABLE_RPC_CONCURRENCY